refactor: separate ToggleList from SingleCardLayout
